Add casts to avoid [s|f]printf() warnings.
Mainly due to use of %u format specifier with formats arguments which are size_t (or promotions to size_t in expressions) which is a typedef of long unsigned int.

+++ SockLibX/ckhentp.cpp
@@ -59,7 +59,7 @@ int NPSL_CheckHostEntParams(const NPSL_HOSTENT *host_ent_ptr,
 		if (error_text != NULL)
 			sprintf(error_text, "%s (%u) is less than the minimum (%u).",
 				"The 'NPSL_HOSTENT' buffer length", host_ent_buffer_length,
-				sizeof(NPSL_HOSTENT));
+				((unsigned int) sizeof(NPSL_HOSTENT)));
 		return_code = NPSL_FAILURE;
 	}
